move CmpOp enum to mod.rs
